chore: build artifacts
