+++ lib/Typesense/ApiCall.js
@@ -35,10 +35,10 @@ var ApiCall = /** @class */ (function () {
         this.initializeMetadataForNodes();
         this.currentNodeIndex = -1;
     }
-    ApiCall.prototype.get = function (endpoint, queryParameters, _a) {
-        if (queryParameters === void 0) { queryParameters = {}; }
-        var _b = _a === void 0 ? {} : _a, _c = _b.abortSignal, abortSignal = _c === void 0 ? null : _c, _d = _b.responseType, responseType = _d === void 0 ? undefined : _d;
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    ApiCall.prototype.get = function (endpoint_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (endpoint, queryParameters, _a) {
+            if (queryParameters === void 0) { queryParameters = {}; }
+            var _b = _a === void 0 ? {} : _a, _c = _b.abortSignal, abortSignal = _c === void 0 ? null : _c, _d = _b.responseType, responseType = _d === void 0 ? undefined : _d;
             return tslib_1.__generator(this, function (_e) {
                 return [2 /*return*/, this.performRequest("get", endpoint, {
                         queryParameters: queryParameters,
@@ -48,19 +48,19 @@ var ApiCall = /** @class */ (function () {
             });
         });
     };
-    ApiCall.prototype.delete = function (endpoint, queryParameters) {
-        if (queryParameters === void 0) { queryParameters = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    ApiCall.prototype.delete = function (endpoint_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (endpoint, queryParameters) {
+            if (queryParameters === void 0) { queryParameters = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.performRequest("delete", endpoint, { queryParameters: queryParameters })];
             });
         });
     };
-    ApiCall.prototype.post = function (endpoint, bodyParameters, queryParameters, additionalHeaders) {
-        if (bodyParameters === void 0) { bodyParameters = {}; }
-        if (queryParameters === void 0) { queryParameters = {}; }
-        if (additionalHeaders === void 0) { additionalHeaders = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    ApiCall.prototype.post = function (endpoint_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (endpoint, bodyParameters, queryParameters, additionalHeaders) {
+            if (bodyParameters === void 0) { bodyParameters = {}; }
+            if (queryParameters === void 0) { queryParameters = {}; }
+            if (additionalHeaders === void 0) { additionalHeaders = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.performRequest("post", endpoint, {
                         queryParameters: queryParameters,
@@ -70,10 +70,10 @@ var ApiCall = /** @class */ (function () {
             });
         });
     };
-    ApiCall.prototype.put = function (endpoint, bodyParameters, queryParameters) {
-        if (bodyParameters === void 0) { bodyParameters = {}; }
-        if (queryParameters === void 0) { queryParameters = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    ApiCall.prototype.put = function (endpoint_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (endpoint, bodyParameters, queryParameters) {
+            if (bodyParameters === void 0) { bodyParameters = {}; }
+            if (queryParameters === void 0) { queryParameters = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.performRequest("put", endpoint, {
                         queryParameters: queryParameters,
@@ -82,10 +82,10 @@ var ApiCall = /** @class */ (function () {
             });
         });
     };
-    ApiCall.prototype.patch = function (endpoint, bodyParameters, queryParameters) {
-        if (bodyParameters === void 0) { bodyParameters = {}; }
-        if (queryParameters === void 0) { queryParameters = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    ApiCall.prototype.patch = function (endpoint_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (endpoint, bodyParameters, queryParameters) {
+            if (bodyParameters === void 0) { bodyParameters = {}; }
+            if (queryParameters === void 0) { queryParameters = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.performRequest("patch", endpoint, {
                         queryParameters: queryParameters,
@@ -105,13 +105,13 @@ var ApiCall = /** @class */ (function () {
             ? axios_1.default.getAdapter(this.configuration.axiosAdapter).bind(globalThis)
             : axios_1.default.getAdapter(this.configuration.axiosAdapter);
     };
-    ApiCall.prototype.performRequest = function (requestType, endpoint, _a) {
-        var _b, _c, _d, _e;
-        var _f = _a.queryParameters, queryParameters = _f === void 0 ? null : _f, _g = _a.bodyParameters, bodyParameters = _g === void 0 ? null : _g, _h = _a.additionalHeaders, additionalHeaders = _h === void 0 ? {} : _h, _j = _a.abortSignal, abortSignal = _j === void 0 ? null : _j, _k = _a.responseType, responseType = _k === void 0 ? undefined : _k, _l = _a.skipConnectionTimeout, skipConnectionTimeout = _l === void 0 ? false : _l, _m = _a.enableKeepAlive, enableKeepAlive = _m === void 0 ? undefined : _m;
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    ApiCall.prototype.performRequest = function (requestType_1, endpoint_1, _a) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (requestType, endpoint, _b) {
             var requestNumber, lastException, wasAborted, _loop_1, this_1, numTries, state_1;
-            return tslib_1.__generator(this, function (_o) {
-                switch (_o.label) {
+            var _c, _d, _e, _f;
+            var _g = _b.queryParameters, queryParameters = _g === void 0 ? null : _g, _h = _b.bodyParameters, bodyParameters = _h === void 0 ? null : _h, _j = _b.additionalHeaders, additionalHeaders = _j === void 0 ? {} : _j, _k = _b.abortSignal, abortSignal = _k === void 0 ? null : _k, _l = _b.responseType, responseType = _l === void 0 ? undefined : _l, _m = _b.skipConnectionTimeout, skipConnectionTimeout = _m === void 0 ? false : _m, _o = _b.enableKeepAlive, enableKeepAlive = _o === void 0 ? undefined : _o;
+            return tslib_1.__generator(this, function (_p) {
+                switch (_p.label) {
                     case 0:
                         this.configuration.validate();
                         requestNumber = Date.now();
@@ -119,8 +119,8 @@ var ApiCall = /** @class */ (function () {
                         this.logger.debug("Request #".concat(requestNumber, ": Performing ").concat(requestType.toUpperCase(), " request: ").concat(endpoint));
                         _loop_1 = function (numTries) {
                             var node, abortListener, requestOptions, cancelToken, source_1, response, error_1;
-                            return tslib_1.__generator(this, function (_p) {
-                                switch (_p.label) {
+                            return tslib_1.__generator(this, function (_q) {
+                                switch (_q.label) {
                                     case 0:
                                         node = this_1.getNextNode(requestNumber);
                                         this_1.logger.debug("Request #".concat(requestNumber, ": Attempting ").concat(requestType.toUpperCase(), " request Try #").concat(numTries, " to Node ").concat(node.index));
@@ -128,9 +128,9 @@ var ApiCall = /** @class */ (function () {
                                             return [2 /*return*/, { value: Promise.reject(new Error("Request aborted by caller.")) }];
                                         }
                                         abortListener = void 0;
-                                        _p.label = 1;
+                                        _q.label = 1;
                                     case 1:
-                                        _p.trys.push([1, 3, 5, 6]);
+                                        _q.trys.push([1, 3, 5, 6]);
                                         requestOptions = {
                                             adapter: this_1.getAdapter(),
                                             method: requestType,
@@ -219,7 +219,7 @@ var ApiCall = /** @class */ (function () {
                                         }
                                         return [4 /*yield*/, (0, axios_1.default)(requestOptions)];
                                     case 2:
-                                        response = _p.sent();
+                                        response = _q.sent();
                                         if (response.status >= 1 && response.status <= 499) {
                                             // Treat any status code > 0 and < 500 to be an indication that node is healthy
                                             // We exclude 0 since some clients return 0 when request fails
@@ -230,24 +230,24 @@ var ApiCall = /** @class */ (function () {
                                             return [2 /*return*/, { value: Promise.resolve(response.data) }];
                                         }
                                         else if (response.status < 500) {
-                                            return [2 /*return*/, { value: Promise.reject(this_1.customErrorForResponse(response, (_b = response.data) === null || _b === void 0 ? void 0 : _b.message, requestOptions.data)) }];
+                                            return [2 /*return*/, { value: Promise.reject(this_1.customErrorForResponse(response, (_c = response.data) === null || _c === void 0 ? void 0 : _c.message, requestOptions.data)) }];
                                         }
                                         else {
                                             // Retry all other HTTP errors (HTTPStatus > 500)
                                             // This will get caught by the catch block below
-                                            throw this_1.customErrorForResponse(response, (_c = response.data) === null || _c === void 0 ? void 0 : _c.message, requestOptions.data);
+                                            throw this_1.customErrorForResponse(response, (_d = response.data) === null || _d === void 0 ? void 0 : _d.message, requestOptions.data);
                                         }
                                         return [3 /*break*/, 6];
                                     case 3:
-                                        error_1 = _p.sent();
+                                        error_1 = _q.sent();
                                         // This block handles retries for HTTPStatus > 500 and network layer issues like connection timeouts
                                         if (!wasAborted) {
                                             this_1.setNodeHealthcheck(node, UNHEALTHY);
                                         }
                                         lastException = error_1;
-                                        this_1.logger.warn("Request #".concat(requestNumber, ": Request to Node ").concat(node.index, " failed due to \"").concat((_d = error_1 === null || error_1 === void 0 ? void 0 : error_1.code) !== null && _d !== void 0 ? _d : "", " ").concat(error_1.message).concat(error_1.response == null
+                                        this_1.logger.warn("Request #".concat(requestNumber, ": Request to Node ").concat(node.index, " failed due to \"").concat((_e = error_1 === null || error_1 === void 0 ? void 0 : error_1.code) !== null && _e !== void 0 ? _e : "", " ").concat(error_1.message).concat(error_1.response == null
                                             ? ""
-                                            : " - " + JSON.stringify((_e = error_1.response) === null || _e === void 0 ? void 0 : _e.data), "\""));
+                                            : " - " + JSON.stringify((_f = error_1.response) === null || _f === void 0 ? void 0 : _f.data), "\""));
                                         // this.logger.debug(error.stack)
                                         if (wasAborted) {
                                             return [2 /*return*/, { value: Promise.reject(new Error("Request aborted by caller.")) }];
@@ -257,7 +257,7 @@ var ApiCall = /** @class */ (function () {
                                         }
                                         return [4 /*yield*/, this_1.timer(this_1.retryIntervalSeconds)];
                                     case 4:
-                                        _p.sent();
+                                        _q.sent();
                                         return [3 /*break*/, 6];
                                     case 5:
                                         if (abortSignal && abortListener) {
@@ -270,15 +270,15 @@ var ApiCall = /** @class */ (function () {
                         };
                         this_1 = this;
                         numTries = 1;
-                        _o.label = 1;
+                        _p.label = 1;
                     case 1:
                         if (!(numTries <= this.numRetriesPerRequest + 1)) return [3 /*break*/, 4];
                         return [5 /*yield**/, _loop_1(numTries)];
                     case 2:
-                        state_1 = _o.sent();
+                        state_1 = _p.sent();
                         if (typeof state_1 === "object")
                             return [2 /*return*/, state_1.value];
-                        _o.label = 3;
+                        _p.label = 3;
                     case 3:
                         numTries++;
                         return [3 /*break*/, 1];
