+++ lib/Typesense/Collection.js
@@ -38,9 +38,9 @@ var Collection = /** @class */ (function () {
             });
         });
     };
-    Collection.prototype.delete = function (options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Collection.prototype.delete = function () {
+        return tslib_1.__awaiter(this, arguments, void 0, function (options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.apiCall.delete(this.endpointPath(), options)];
             });
