+++ lib/Typesense/Collections.js
@@ -13,9 +13,9 @@ var Collections = /** @class */ (function () {
             });
         });
     };
-    Collections.prototype.retrieve = function (options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Collections.prototype.retrieve = function () {
+        return tslib_1.__awaiter(this, arguments, void 0, function (options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.apiCall.get(RESOURCEPATH, options)];
             });
