+++ lib/Typesense/Document.js
@@ -24,9 +24,9 @@ var Document = /** @class */ (function () {
             });
         });
     };
-    Document.prototype.update = function (partialDocument, options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Document.prototype.update = function (partialDocument_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (partialDocument, options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.apiCall.patch(this.endpointPath(), partialDocument, options)];
             });
