+++ lib/Typesense/Documents.js
@@ -1,26 +1,8 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.arrayableParams = void 0;
 var tslib_1 = require("tslib");
 var Errors_1 = require("./Errors");
 var SearchOnlyDocuments_1 = require("./SearchOnlyDocuments");
-exports.arrayableParams = {
-    query_by: "query_by",
-    query_by_weights: "query_by_weights",
-    facet_by: "facet_by",
-    group_by: "group_by",
-    include_fields: "include_fields",
-    exclude_fields: "exclude_fields",
-    highlight_fields: "highlight_fields",
-    highlight_full_fields: "highlight_full_fields",
-    pinned_hits: "pinned_hits",
-    hidden_hits: "hidden_hits",
-    infix: "infix",
-    override_tags: "override_tags",
-    num_typos: "num_typos",
-    prefix: "prefix",
-    sort_by: "sort_by",
-};
 var isNodeJSEnvironment = typeof process !== "undefined" &&
     process.versions != null &&
     process.versions.node != null;
@@ -29,9 +11,9 @@ var Documents = /** @class */ (function (_super) {
     function Documents(collectionName, apiCall, configuration) {
         return _super.call(this, collectionName, apiCall, configuration) || this;
     }
-    Documents.prototype.create = function (document, options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.create = function (document_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (document, options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 if (!document)
                     throw new Error("No document provided");
@@ -39,9 +21,9 @@ var Documents = /** @class */ (function (_super) {
             });
         });
     };
-    Documents.prototype.upsert = function (document, options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.upsert = function (document_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (document, options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 if (!document)
                     throw new Error("No document provided");
@@ -49,9 +31,9 @@ var Documents = /** @class */ (function (_super) {
             });
         });
     };
-    Documents.prototype.update = function (document, options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.update = function (document_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (document, options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 if (!document)
                     throw new Error("No document provided");
@@ -65,27 +47,27 @@ var Documents = /** @class */ (function (_super) {
             });
         });
     };
-    Documents.prototype.delete = function (query) {
-        if (query === void 0) { query = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.delete = function () {
+        return tslib_1.__awaiter(this, arguments, void 0, function (query) {
+            if (query === void 0) { query = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.apiCall.delete(this.endpointPath(), query)];
             });
         });
     };
-    Documents.prototype.createMany = function (documents, options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.createMany = function (documents_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (documents, options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 this.configuration.logger.warn("createMany is deprecated and will be removed in a future version. Use import instead, which now takes both an array of documents or a JSONL string of documents");
                 return [2 /*return*/, this.import(documents, options)];
             });
         });
     };
-    Documents.prototype.import = function (documents, options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.import = function (documents_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (documents, options) {
             var documentsInJSONLFormat, resultsInJSONLFormat, resultsInJSONFormat, failedItems;
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 switch (_a.label) {
                     case 0:
@@ -151,10 +133,10 @@ var Documents = /** @class */ (function (_super) {
     /**
      * Imports documents from a NodeJS readable stream of JSONL.
      */
-    Documents.prototype.importStream = function (readableStream, options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.importStream = function (readableStream_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (readableStream, options) {
             var resultsInJSONLFormat, resultsInJSONFormat, failedItems;
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 switch (_a.label) {
                     case 0: return [4 /*yield*/, this.apiCall.performRequest("post", this.endpointPath("import"), {
@@ -189,9 +171,9 @@ var Documents = /** @class */ (function (_super) {
     /**
      * Returns a JSONL string for all the documents in this collection
      */
-    Documents.prototype.export = function (options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.export = function () {
+        return tslib_1.__awaiter(this, arguments, void 0, function (options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.apiCall.get(this.endpointPath("export"), options)];
             });
@@ -200,9 +182,9 @@ var Documents = /** @class */ (function (_super) {
     /**
      * Returns a NodeJS readable stream of JSONL for all the documents in this collection.
      */
-    Documents.prototype.exportStream = function (options) {
-        if (options === void 0) { options = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Documents.prototype.exportStream = function () {
+        return tslib_1.__awaiter(this, arguments, void 0, function (options) {
+            if (options === void 0) { options = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.apiCall.get(this.endpointPath("export"), options, {
                         responseType: "stream",
