+++ lib/Typesense/Errors/ImportError.d.ts
@@ -1,4 +1,3 @@
-/// <reference types="node" />
 import TypesenseError from "./TypesenseError";
 import type { DocumentImportParameters, ImportResponse } from "../Documents";
 import { ReadStream } from "node:fs";
