+++ lib/Typesense/MultiSearch.js
@@ -15,12 +15,12 @@ var MultiSearch = /** @class */ (function () {
     MultiSearch.prototype.clearCache = function () {
         this.requestWithCache.clearCache();
     };
-    MultiSearch.prototype.perform = function (searchRequests, commonParams, _a) {
-        if (commonParams === void 0) { commonParams = {}; }
-        var _b = _a === void 0 ? {} : _a, _c = _b.cacheSearchResultsForSeconds, cacheSearchResultsForSeconds = _c === void 0 ? this.configuration
-            .cacheSearchResultsForSeconds : _c;
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    MultiSearch.prototype.perform = function (searchRequests_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (searchRequests, commonParams, _a) {
             var additionalHeaders, additionalQueryParams, queryParams, normalizedSearchRequests, normalizedQueryParams;
+            if (commonParams === void 0) { commonParams = {}; }
+            var _b = _a === void 0 ? {} : _a, _c = _b.cacheSearchResultsForSeconds, cacheSearchResultsForSeconds = _c === void 0 ? this.configuration
+                .cacheSearchResultsForSeconds : _c;
             return tslib_1.__generator(this, function (_d) {
                 additionalHeaders = {};
                 if (this.useTextContentType) {
