+++ lib/Typesense/Operations.js
@@ -6,9 +6,9 @@ var Operations = /** @class */ (function () {
     function Operations(apiCall) {
         this.apiCall = apiCall;
     }
-    Operations.prototype.perform = function (operationName, queryParameters) {
-        if (queryParameters === void 0) { queryParameters = {}; }
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    Operations.prototype.perform = function (operationName_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (operationName, queryParameters) {
+            if (queryParameters === void 0) { queryParameters = {}; }
             return tslib_1.__generator(this, function (_a) {
                 return [2 /*return*/, this.apiCall.post("".concat(RESOURCEPATH, "/").concat(operationName), {}, queryParameters)];
             });
