+++ lib/Typesense/SearchOnlyDocuments.js
@@ -16,11 +16,11 @@ var SearchOnlyDocuments = /** @class */ (function () {
     SearchOnlyDocuments.prototype.clearCache = function () {
         this.requestWithCache.clearCache();
     };
-    SearchOnlyDocuments.prototype.search = function (searchParameters, _a) {
-        var _b = _a === void 0 ? {} : _a, _c = _b.cacheSearchResultsForSeconds, cacheSearchResultsForSeconds = _c === void 0 ? this.configuration
-            .cacheSearchResultsForSeconds : _c, _d = _b.abortSignal, abortSignal = _d === void 0 ? null : _d;
-        return tslib_1.__awaiter(this, void 0, void 0, function () {
+    SearchOnlyDocuments.prototype.search = function (searchParameters_1) {
+        return tslib_1.__awaiter(this, arguments, void 0, function (searchParameters, _a) {
             var additionalQueryParams, normalizedParams, queryParams;
+            var _b = _a === void 0 ? {} : _a, _c = _b.cacheSearchResultsForSeconds, cacheSearchResultsForSeconds = _c === void 0 ? this.configuration
+                .cacheSearchResultsForSeconds : _c, _d = _b.abortSignal, abortSignal = _d === void 0 ? null : _d;
             return tslib_1.__generator(this, function (_e) {
                 additionalQueryParams = {};
                 if (this.configuration.useServerSideSearchCache === true) {
