+++ lib/Typesense/Types.js
@@ -1,0 +1,21 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.arrayableParams = void 0;
+exports.arrayableParams = {
+    query_by: "query_by",
+    query_by_weights: "query_by_weights",
+    facet_by: "facet_by",
+    group_by: "group_by",
+    include_fields: "include_fields",
+    exclude_fields: "exclude_fields",
+    highlight_fields: "highlight_fields",
+    highlight_full_fields: "highlight_full_fields",
+    pinned_hits: "pinned_hits",
+    hidden_hits: "hidden_hits",
+    infix: "infix",
+    override_tags: "override_tags",
+    num_typos: "num_typos",
+    prefix: "prefix",
+    sort_by: "sort_by",
+};
+//# sourceMappingURL=Types.js.map
