+++ lib/Typesense/Utils.d.ts
@@ -1,4 +1,4 @@
-import type { ExtractBaseTypes, SearchParams } from "./Documents";
+import type { ExtractBaseTypes, SearchParams } from "./Types";
 export declare function normalizeArrayableParams<T extends SearchParams>(params: T): Prettify<ExtractBaseTypes<T>>;
 type Prettify<T> = {
     [K in keyof T]: T[K];
