+++ lib/Typesense/Utils.js
@@ -1,16 +1,16 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.normalizeArrayableParams = void 0;
+exports.normalizeArrayableParams = normalizeArrayableParams;
 var tslib_1 = require("tslib");
-var Documents_1 = require("./Documents");
+var Types_1 = require("./Types");
 function hasNoArrayValues(params) {
-    return Object.keys(Documents_1.arrayableParams)
+    return Object.keys(Types_1.arrayableParams)
         .filter(function (key) { return params[key] !== undefined; })
         .every(function (key) { return isNonArrayValue(params[key]); });
 }
 function normalizeArrayableParams(params) {
     var result = tslib_1.__assign({}, params);
-    var transformedValues = Object.keys(Documents_1.arrayableParams)
+    var transformedValues = Object.keys(Types_1.arrayableParams)
         .filter(function (key) { return Array.isArray(result[key]); })
         .map(function (key) {
         result[key] = result[key].join(",");
@@ -24,7 +24,6 @@ function normalizeArrayableParams(params) {
     }
     return result;
 }
-exports.normalizeArrayableParams = normalizeArrayableParams;
 function isNonArrayValue(value) {
     return !Array.isArray(value);
 }
